fix scope equality check
