+++ frontend/openchat-client/src/state/global/index.ts
@@ -1,2 +1,1 @@
-export { global } from "./global.svelte";
 export { globalLocalUpdates } from "./local.svelte";
